feat: support dev portal link oauth2 account redirect (#29)
Added an optional `after_oauth2_link_redirect_url` query param to the
`GET /v1/linked-accounts/oauth2` endpoint. If this param is provided, it
will be encoded into the `state` of the OAuth flow. The `GET
/v1/linked-accounts/oauth2/callback` will look for this field in
`state`. If it exists, redirect to the URL specified, otherwise, return
a JSON response as normal.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Introduced an option for automatic redirection after successful
account linking, allowing for a smoother OAuth2 experience.
- Enhanced API interactions by now supporting both GET and POST
requests, improving how data is exchanged between clients and the
server.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->