feat: allow creating and editing agents (#18)
### 📝 Description

* Commented out `ENABLED APPS` column as it's out of the scope of this
release
* Allow users to create new agents
* Allow users to edit the name, description, and custom_instructions of
agents

### 🎥 Demo

Two things about this demo:

1. You see an error in the bottom left corner, this is because the
`react-json-view` library is not fully compatible with `react-hook-form`
library. This error is because some attributes set on an element is
ignored, which was shown as an error by the Next.js development server.
Users won't see it in prod and it won't affect the functionality.
We'll get rid of this form once we move to a proper agent edit UI anyway
so not worth fixing it now.

2. I made the default value for custom instructions to be empty strings
after I recorded the demo, so users won't see a `null` when they added a
new key in custom instructions.


https://github.com/user-attachments/assets/20ca4254-adfd-4cf5-ab0f-2d8ad1eab9c9



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Enhanced project settings with integrated agent management, allowing
users to create and edit agents using a streamlined form with inline
validation and clear presentation of custom instructions.
- Introduced a versatile multi-line text input component for improved
data entry.
- Added new asynchronous functions for creating and updating agents,
enhancing data handling capabilities.
- **Bug Fixes**
- Improved project navigation by ensuring the selector menu closes
automatically when accessing project settings.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->