Feat/agent level app access control (#87)
### 🏷️ Notion Ticket


[https://www.notion.so/Allow-agent-level-App-access-1b28378d6a4780f3b66bc5fb3064f8c3]

### 📝 Description
Support agent level app access control: agent now have `allowed_apps`
fields (we used to have `excluded_apps`), which should be a subset of
configured apps under the project. When executing functions, we should
check if the function belongs to the app that is outside the agent's
(identified by api key) allowed apps. If so return 401.

This field is also used during app and functions search if the search is
specified with `allowed_apps_ony` (replacing the previously
`configured_only` flag) to only search apps and functions accessible by
requesting agent.

Updated the cli command `create_random_api_key` to include all apps as
allowed apps for the sake of easier local testing. Otherwise the api_key
(of the agent) won't be able to used to execute functions.